Moved passive joint class into header.
